// opencdx-dashboard-new/src/api/iam/model/workspace.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { Department } from './department';
// May contain unused imports in some cases
// @ts-ignore
import type { Timestamp } from './timestamp';

/**
 * 
 * @export
 * @interface Workspace
 */
export interface Workspace {
    /**
     * 
     * @type {string}
     * @memberof Workspace
     */
    'id'?: string;
    /**
     * 
     * @type {string}
     * @memberof Workspace
     */
    'name'?: string;
    /**
     * 
     * @type {string}
     * @memberof Workspace
     */
    'description'?: string;
    /**
     * 
     * @type {string}
     * @memberof Workspace
     */
    'organizationId'?: string;
    /**
     * 
     * @type {Timestamp}
     * @memberof Workspace
     */
    'createdDate'?: Timestamp;
    /**
     * 
     * @type {string}
     * @memberof Workspace
     */
    'location'?: string;
    /**
     * 
     * @type {string}
     * @memberof Workspace
     */
    'manager'?: string;
    /**
     * 
     * @type {number}
     * @memberof Workspace
     */
    'capacity'?: number;
    /**
     * 
     * @type {Array<string>}
     * @memberof Workspace
     */
    'facilities'?: Array<string>;
    /**
     * 
     * @type {string}
     * @memberof Workspace
     */
    'workspaceType'?: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof Workspace
     */
    'workspaceImageUrls'?: Array<string>;
    /**
     * 
     * @type {string}
     * @memberof Workspace
     */
    'usagePolicy'?: string;
    /**
     * 
     * @type {string}
     * @memberof Workspace
     */
    'availabilitySchedule'?: string;
    /**
     * 
     * @type {Array<Department>}
     * @memberof Workspace
     */
    'departments'?: Array<Department>;
    /**
     * 
     * @type {Timestamp}
     * @memberof Workspace
     */
    'created'?: Timestamp;
    /**
     * 
     * @type {Timestamp}
     * @memberof Workspace
     */
    'modified'?: Timestamp;
    /**
     * 
     * @type {string}
     * @memberof Workspace
     */
    'creator'?: string;
    /**
     * 
     * @type {string}
     * @memberof Workspace
     */
    'modifier'?: string;
}